Import libhidx into the repository instead of using a submodule
My dream was to have libhidx as a standalone library that can be used
with other projects as well. Sadly, this has never become true and the
submodule setup became just a hurdle. Let's drop the submodule and just
put the library directly into the repository. This should simplify
any maintanance a bit.